Initial commit
First public commit of Code Master.

Code Master will be acting under the No License terms as stated here:
http://choosealicense.com/no-license/

As for now, Code Master will not be allowing the use of distribution
for any use, whether it be commercial or personal. Changes may be made
to the licensing in the future; however, as of right now, I have not a
chance to fully decide the future of Code Master and its contents.

Code Master 5 is heavily based off the Edhita project
(https://github.com/tnantoka/edhita) acting under the MIT license
(https://github.com/tnantoka/edhita/blob/master/LICENSE).